Fix FSR jitter contract and temporal export quality
- pass the exact camera principal-point jitter to FidelityFX FSR3 so export motion-vector cancellation matches the rendered frame

- align temporal Performance and Quality reconstruction parameters with their visible image-quality semantics

- make 16-bit splat precision the standard default across video options, Sequencer state, events, and regression coverage

- add focused regression tests for FSR jitter deltas, temporal preset ordering, and export precision defaults

+++ src/io/include/io/video/video_export_options.hpp
@@ -82,7 +82,7 @@ namespace lfs::io::video {
         int framerate = 30;
         int crf = 18;
         VideoUpscalerOptions upscaler{};
-        VideoSplatPrecision splat_precision = VideoSplatPrecision::Float32;
+        VideoSplatPrecision splat_precision = VideoSplatPrecision::Float16;
     };
 
     [[nodiscard]] inline std::expected<void, std::string> validateVideoEncodingOptions(
