+++ src/visualizer/gui/sequencer_ui_state.hpp
@@ -28,7 +28,7 @@ namespace lfs::vis::gui::panels {
         float export_input_scale = 1.0f;
         int export_upscaler_quality = 1;
         lfs::io::video::VideoSplatPrecision export_splat_precision =
-            lfs::io::video::VideoSplatPrecision::Float32;
+            lfs::io::video::VideoSplatPrecision::Float16;
     };
 
 } // namespace lfs::vis::gui::panels
